change to `python -m db_server`
